dodano pole aktyny user, korekty w obliczeniach

+++ piping/app/controllers/UserEditCtrl.class.php
@@ -12,6 +12,8 @@ class UserEditCtrl {
 
     private $form;
     private $roles;
+    private $admins;
+    private $actives;
 
     public function __construct() {
         $this->form = new UserEditForm();
@@ -22,6 +24,7 @@ class UserEditCtrl {
         $this->form->login = ParamUtils::getFromRequest('login', true, 'Błędne wywołanie aplikacji');
         $this->form->password = ParamUtils::getFromRequest('password', true, 'Błędne wywołanie aplikacji');
         $this->form->role = ParamUtils::getFromRequest('role', true, 'Błędne wywołanie aplikacji');
+        $this->form->active = ParamUtils::getFromRequest('active', true, 'Błędne wywołanie aplikacji');
 
         if (App::getMessages()->isError())
             return false;
@@ -34,6 +37,9 @@ class UserEditCtrl {
         }
         if (empty(trim($this->form->role))) {
             Utils::addErrorMessage('Wprowadź role');
+        }
+        if (empty(trim($this->form->active))) {
+            Utils::addErrorMessage('Wprowadź aktywność konta');
         }
 
         if (App::getMessages()->isError())
@@ -64,13 +70,18 @@ class UserEditCtrl {
         $this->generateView();
     }
 
-    public function getRolesForm() {
+    public function getRolesActiveForm() {
         try {
             $this->roles = App::getDB()->select("roles", [
                 "idroles",
                 "roles",
                 ]
             );
+            $this->actives = App::getDB()->select("users", [
+                "active",
+                ], [
+                "idusers" => $this->form->id
+            ]);
         } catch (\PDOException $e) {
             Utils::addErrorMessage('Wystąpił nieoczekiwany błąd podczas zapisu rekordu');
             if (App::getConf()->debug)
@@ -88,30 +99,54 @@ class UserEditCtrl {
                 $this->form->login = $record['login'];
                 $this->form->password = $record['password'];
                 $this->form->role = $record['idroles'];
+                $this->form->active = $record['active'];
             } catch (\PDOException $e) {
                 Utils::addErrorMessage('Wystąpił błąd podczas odczytu rekordu');
                 if (App::getConf()->debug)
                     Utils::addErrorMessage($e->getMessage());
             }
-        }
+        } 
 
         $this->generateView();
     }
 
     public function action_userDelete() {
+        $this->lastAdmin();
         if ($this->validateEdit()) {
-            try {
-                App::getDB()->delete("users", [
-                    "idusers" => $this->form->id
-                ]);
-                Utils::addInfoMessage('Pomyślnie usunięto rekord');
-            } catch (\PDOException $e) {
-                Utils::addErrorMessage('Wystąpił błąd podczas usuwania rekordu');
-                if (App::getConf()->debug)
-                    Utils::addErrorMessage($e->getMessage());
+            if($this->admins > 2) {
+                try {
+                    App::getDB()->delete("users", [
+                        "idusers" => $this->form->id
+                    ]);
+                    Utils::addInfoMessage('Pomyślnie usunięto rekord');
+                } catch (\PDOException $e) {
+                    Utils::addErrorMessage('Wystąpił błąd podczas usuwania rekordu');
+                    if (App::getConf()->debug)
+                        Utils::addErrorMessage($e->getMessage());
+                }
+            } else {
+                Utils::addErrorMessage('W systemie musi pozostać 2-óch administratorów');
             }
         }
+
         App::getRouter()->forwardTo('userList');
+    }
+
+    public function lastAdmin() {
+        try {
+            $this->admins = App::getDB()->count("users", [
+                "[>]roles" => "idroles"
+            ], [
+                "idroles"
+            ], [
+                "roles" => 'admin'
+            ]
+            );
+        } catch (\PDOException $e) {
+            Utils::addErrorMessage('Wystąpił nieoczekiwany błąd podczas zapisu rekordu');
+            if (App::getConf()->debug)
+                Utils::addErrorMessage($e->getMessage());
+        }
     }
 
     public function action_userSave() {
@@ -121,14 +156,16 @@ class UserEditCtrl {
                     App::getDB()->insert("users", [
                         "login" => $this->form->login,
                         "password" => $this->form->password,
-                        "idroles" => $this->form->role
+                        "idroles" => $this->form->role,
+                        "active" => $this->form->active
                     ]);
                     Utils::addInfoMessage('Pomyślnie zapisano rekord');
                 } else {
                     App::getDB()->update("users", [
                         "login" => $this->form->login,
                         "password" => $this->form->password,
-                        "idroles" => $this->form->role
+                        "idroles" => $this->form->role,
+                        "active" => $this->form->active
                         ], [
                         "idusers" => $this->form->id
                         ]);
@@ -145,9 +182,12 @@ class UserEditCtrl {
     }
 
     public function generateView() {
-        $this->getRolesForm();
-        // print_r($this->roles);
+        $this->getRolesActiveForm();
+        $this->lastAdmin();
+        // print_r($this->actives);
+        // print_r($this->admins);
         App::getSmarty()->assign('roles', $this->roles);
+        App::getSmarty()->assign('actives', $this->actives);
         App::getSmarty()->assign('form', $this->form);
         App::getSmarty()->display('userNew.tpl');
     }
